# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Nekretnine API",
        "description": "API dokumentacija za CRM sistem nekretnina",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ponude": {
            "get": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Lista ponuda",
                "operationId": "b528624c0de596e8a685a2cc9660736f",
                "responses": {
                    "200": {
                        "description": "Lista ponuda"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Kreiranje ponude",
                "operationId": "980b2e183efa678f579dfdbace0cf083",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kupac_id",
                                    "nekretnina_id",
                                    "cena",
                                    "datum"
                                ],
                                "properties": {
                                    "kupac_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nekretnina_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "cena": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "u_toku"
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Kupac nudi 120k, traži brzu realizaciju."
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ponuda kreirana"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ponude/{id}": {
            "get": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Prikaz ponude po ID-u",
                "operationId": "c76d25874e7dfe3ecb68e0b4e8d251bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ponude",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji ponude"
                    },
                    "404": {
                        "description": "Ponuda nije pronađena"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Izmena ponude",
                "operationId": "b7ded6ddf1106b70cd0dfc84c8c3ef06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ponude",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kupac_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nekretnina_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "cena": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 125000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "prihvacena"
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Dogovoreno na 125k."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ponuda izmenjena"
                    },
                    "404": {
                        "description": "Ponuda nije pronađena"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Brisanje ponude",
                "operationId": "56385728c698a44d9c7adb15dc07e28a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ponude",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ponuda obrisana"
                    },
                    "404": {
                        "description": "Ponuda nije pronađena"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ponude/search": {
            "get": {
                "tags": [
                    "Ponude"
                ],
                "summary": "Pretraga ponuda",
                "operationId": "290884f5bc8a1c3c2c513b5346681d74",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Tekst pretrage (status/napomena)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "u_toku"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Broj strane",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj stavki po strani",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter po statusu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "u_toku"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Polje za sortiranje",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "cena",
                                "status",
                                "kupac_id",
                                "nekretnina_id"
                            ]
                        },
                        "example": "created_at"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Smer sortiranja",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginirani rezultat pretrage"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}